feat: refactor component imports and add HeaderRightMenu, Footer, and Zoom buttons
