+++ frontend/src/features/prototype/components/atoms/ZoomInButton.tsx
@@ -1,6 +1,6 @@
 import { AiOutlineZoomIn } from 'react-icons/ai';
 
-import IconButton from './IconButton';
+import IconButton from '../../../../components/atoms/IconButton';
 
 export default function ZoomInButton({
   onClick,
